+++ CHANGES.rst
@@ -21,6 +21,9 @@ New Features
 
 - Hitting the "Enter" key while renaming a subset or footprint will accept the changes, hitting "Escape" will cancel. [#3600]
 
+- Added subset_label keyword argument to ``import_region`` method of Subset Tools plugin
+  to name the resulting subset(s). [#3616]
+
 Cubeviz
 ^^^^^^^